+++ lib/karafka/pro/processing/strategies/lrj_mom.rb
@@ -50,7 +50,7 @@ module Karafka
 
                 resume
               else
-                pause(coordinator.seek_offset, false)
+                retry_after_pause
               end
             end
           end